move from pipenv to poetry
after some long testing and reading in google i realize that using pipenv might
be outdated and now poetry is the best thing to use. it uses the PEP 518
formatting standards and has some other things that are nice. above all, it
seems my editor of spacemacs is able to recognize its paths and virtualenvs
better than pipenv at this time