some what there...
trying to get a list of possible moves for a selected piece but having
issues with finding the next move based on if the player is black or
white. maybe i should look at a simple test and work my way towrds the
finish with that. also, the check_space_status should be broken out into
individual functions because its too big right now.